pyproject.toml
[tool.poetry]
name = "mobileapi"
version = "0.1.0"
description = ""
authors = ["bayan-alkhatib <bayankhatibtr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.1"
djangorestframework-simplejwt = "^4.7.2"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.1"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
